backend/requirements.txt
# Backend dependencies
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pymongo>=4.0.0
werkzeug>=3.0.0

